Calculates FOTO classification of texture for an image batch

This routine process images as a batch, normalizing the PCA analysis across images. This global normalization makes it possible to compare the resulting PCA scores across images and infer trends over different remote sensing tiles or across time.

foto_batch(path, window_size = 61, method = "zones", cores = 1)

Arguments

path

directory containing (only) image files to process

window_size

a moving window size in pixels (default = 61 pixels)

method

zones (for discrete zones) or mw for a moving window approach

cores

number of cores to use in parallel calculations

Value

returns a radial spectrum for a moving window across a raster layer

See also

Examples

if (FALSE) {
# load demo data path
path <- system.file("extdata", package = "foto")

# classify pixels using zones (discrete steps)
output <- foto_batch(
  path = path,
  window_size = 25,
  method = "zones"
)
}